+++ src/components/seller/index.ts
@@ -8,3 +8,8 @@ export { SellerTopProducts } from "./SellerTopProducts";
 export type { SellerAnalyticsSummary } from "./SellerAnalyticsStats";
 export type { MonthEntry } from "./SellerRevenueChart";
 export type { TopProduct } from "./SellerTopProducts";
+export { SellerPayoutStats } from "./SellerPayoutStats";
+export { SellerPayoutRequestForm } from "./SellerPayoutRequestForm";
+export { SellerPayoutHistoryTable } from "./SellerPayoutHistoryTable";
+export type { PayoutSummary } from "./SellerPayoutStats";
+export type { PayoutRecord } from "./SellerPayoutHistoryTable";